+++ docs/changelog.rst
@@ -11,6 +11,9 @@ API Changes
 
 New Features
 ~~~~~~~~~~~~
+- Added functions to compute error bars for transition probabilities to account for
+  finite sampling, and sample transition matrices from these error distributions (i.e.
+  bootstrapping). Located in ```msmbuilder.msm.validation.transmat_errorbar```.
 
 Improvements
 ~~~~~~~~~~~~